+++ sipp_node.h
@@ -8,7 +8,7 @@ struct SIPPNode : virtual public Node
     int     startTime;
     int     endTime;
 
-    SIPPNode(int x = 0, int y = 0, Node *p = nullptr, int g_ = 0, int H_ = 0, int ConflictsCount = 0, int hc_ = 0) :
+    SIPPNode(int x = 0, int y = 0, Node *p = nullptr, int g_ = 0, double H_ = 0, int ConflictsCount = 0, int hc_ = 0) :
         Node(x, y, p, g_, H_, ConflictsCount), startTime(g_), endTime(g_) {}
 
     SIPPNode(const Node &other) : Node(other) {}
